requirements.txt
selenium
webdriver-manager
Pillow
requests
numpy